De-dup with TorchMultiprocessQueueFactory
