+++ tsercom/api/split_process/split_runtime_factory_factory.py
@@ -27,7 +27,7 @@ from tsercom.threading.multiprocess.multiprocess_queue_factory import (
     MultiprocessQueueFactory,
 )
 from tsercom.threading.multiprocess.torch_multiprocess_queue_factory import (
-    TorchMultiprocessQueueFactory,
+    TorchMemcpyQueueFactory,
 )
 from tsercom.threading.multiprocess.multiprocess_queue_sink import (
     MultiprocessQueueSink,
@@ -136,10 +136,10 @@ class SplitRuntimeFactoryFactory(RuntimeFactoryFactory[DataTypeT, EventTypeT]):
 
         if uses_torch_tensor:
             # Assuming EventInstance and AnnotatedInstance generics are compatible with Torch queues
-            event_queue_factory = TorchMultiprocessQueueFactory[
+            event_queue_factory = TorchMemcpyQueueFactory[
                 EventInstance[EventTypeT]
             ]()
-            data_queue_factory = TorchMultiprocessQueueFactory[
+            data_queue_factory = TorchMemcpyQueueFactory[
                 AnnotatedInstance[DataTypeT]
             ]()
         else:
